+++ bender/bender-base/app/src/main/jni/mesh.h
@@ -31,7 +31,7 @@ public:
   Mesh(Renderer& renderer, std::shared_ptr<Material> material, std::shared_ptr<Geometry> geometryData);
 
   Mesh(Renderer& renderer, std::shared_ptr<Material> material, const std::vector<float>& vertexData,
-          const std::vector<uint16_t>& indexData);
+          const std::vector<uint32_t>& indexData);
 
   ~Mesh();
 
@@ -41,7 +41,7 @@ public:
   void update(uint_t frame_index, glm::vec3 camera, glm::mat4 view, glm::mat4 proj);
   void submitDraw(VkCommandBuffer commandBuffer, uint_t frame_index) const;
 
-  void swapGeometry(const std::vector<float>& vertexData, const std::vector<uint16_t>& indexData);
+  void swapGeometry(const std::vector<float>& vertexData, const std::vector<uint32_t>& indexData);
   void swapMaterial(std::shared_ptr<Material> material) { material_ = material; }
 
   void translate(glm::vec3 offset);
